+++ app/loading.tsx
@@ -1,5 +1,16 @@
-import Loading from './_components/Loading'
+'use client'
+import Lottie from 'react-lottie-player'
+import loadingAnimation from '@/public/animations/loading.json'
 
-export default function loading() {
-  return <Loading />
+export default function Loading() {
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center bg-white">
+      <Lottie
+        loop
+        animationData={loadingAnimation}
+        play
+        style={{ width: 500, height: 500 }}
+      />
+    </div>
+  )
 }
